+++ packages/harmony/src/components/input/PasswordInput/PasswordInput.modules.css
@@ -4,10 +4,17 @@
   display: flex;
   color: var(--harmony-text-subdued);
   transition: border ease-in-out 0.1s;
+  border-radius: 50%;
+  padding: var(--harmony-spacing-xs);
 }
 
 .button:hover {
   color: var(--harmony-n-300);
+  background-color: var(--harmony-n-100);
+}
+
+.button:active {
+  background-color: var(--harmony-n-150);
 }
 
 .button svg {
