+++ packages/web/src/components/form-fields/HarmonyTextField.tsx
@@ -1,6 +1,5 @@
-import { TextInput } from '@audius/harmony'
-import { TextInputProps } from '@audius/harmony/dist/components/input/TextInput/types'
-import { HarmonyBaseTextField } from './HarmonyBaseTextField'
+import { TextInput, TextInputProps } from '@audius/harmony'
+import { useField } from 'formik'
 
 export type TextFieldProps = TextInputProps & {
   name: string
@@ -13,10 +12,23 @@ export type TextFieldProps = TextInputProps & {
 
 // TODO: rename to TextField and replace old usages
 export const HarmonyTextField = (props: TextFieldProps) => {
+  const { name, clearErrorOnChange = true, ...other } = props
+  const [field, { touched, error }, { setError }] = useField(name)
+
+  const hasError = Boolean(touched && error)
+
   return (
-    <HarmonyBaseTextField<TextInputProps>
-      inputComponent={TextInput}
-      {...props}
+    <TextInput
+      {...field}
+      error={hasError}
+      helperText={hasError ? error : undefined}
+      onChange={(e) => {
+        if (clearErrorOnChange) {
+          setError(undefined)
+        }
+        field.onChange(e)
+      }}
+      {...other}
     />
   )
 }
