+++ packages/web/src/pages/sign-up-page/styles/CreatePasswordPage.module.css
@@ -9,7 +9,3 @@
 .backIcon:hover {
   color: var(--harmony-n-600);
 }
-
-.backIcon path {
-  fill: currentColor;
-}
